libstagefright: Make it possible to skip OMX buffer reallocation
Some devices don't like the call to setParameter() at this point, so
skip this call if enough buffers are already allocated. This check
was present in KitKat but got removed when code to allocate extra-
buffers was introduced.

This is activated only for omap4 for now.

Squashed 2nd commit: which does same for OMXCodec.cpp

Change-Id: I9c1267c2a2a755bc16448a6049cb2e4aff0b65bb
Signed-off-by: Josue Rivera <prbassplayer@gmail.com>