libstagefright: Add support for NV21 color format
This adds the native color format 'OMX_SEC_COLOR_FormatNV21Linear' of
exynos5 which is NV21 and needed for the camera. You can turn it on
with:

BOARD_USE_SAMSUNG_COLORFORMAT_NV21 := true

It requires

const char CameraParameters::PIXEL_FORMAT_YUV420SP_NV21[] = "nv21";

to be specified in CameraExtraParameters.h.

Change-Id: I64ca86d074468b1e7a2958a73c3c0380e64a9fc0
Signed-off-by: Andreas Schneider <asn@cryptomilk.org>