MediaExtractor: Recreate the extractor after fallback detection
This is a workaround for the following CTS tests:
android.media.cts.MediaMetadataRetrieverTest#testRetrieveFailsIfMediaDataSourceReturnsAnError
android.media.cts.MediaMetadataRetrieverTest#testRetrieveFailsIfMediaDataSourceThrows

These tests assume that extractors are lazy and  won't analyze
the source until a getter that depends on metadata is called.

Our method to fallback to ffmpeg (no tracks present or less than
two tracks on video streams) relies on reading metadata to pick
the right extractor which makes it non-lazy.

The MetadataRetriever API does not define whether this assumption
should be enforced but it is conceivable that real-world callers
have the same expectation as the test.

Change-Id: I18b5513c6033ef01f04ed14dec5876ff1be046db